EECS498-A1
This repository contains my implementation for Assignment 1, which serves as an introduction to PyTorch and the fundamentals of image classification. The project involves building a k-Nearest Neighbor (kNN) classifier from scratch, focusing on data-driven approaches, hyperparameter tuning using validation sets, and integration with Google Colab and an Autograder for evaluation.